ScheduleStatic: put in separate file
