+++ opm/input/eclipse/Schedule/Schedule.hpp
@@ -32,14 +32,12 @@
 #include <utility>
 #include <vector>
 
-#include <opm/input/eclipse/EclipseState/Runspec.hpp>
 #include <opm/input/eclipse/Schedule/Action/WGNames.hpp>
 #include <opm/input/eclipse/Schedule/CompletedCells.hpp>
 #include <opm/input/eclipse/Schedule/Group/Group.hpp>
-#include <opm/input/eclipse/Schedule/MessageLimits.hpp>
 #include <opm/input/eclipse/Schedule/ScheduleDeck.hpp>
 #include <opm/input/eclipse/Schedule/ScheduleState.hpp>
-#include <opm/input/eclipse/Schedule/ScheduleRestartInfo.hpp>
+#include <opm/input/eclipse/Schedule/ScheduleStatic.hpp>
 #include <opm/input/eclipse/Schedule/Well/Connection.hpp>
 #include <opm/input/eclipse/Schedule/Well/PAvg.hpp>
 #include <opm/input/eclipse/Schedule/WriteRestartFileEvents.hpp>
@@ -68,6 +66,7 @@ namespace Opm
     enum class InputErrorAction;
     class ParseContext;
     class Python;
+    class Runspec;
     class RPTConfig;
     class ScheduleGrid;
     class SCHEDULESection;
@@ -85,56 +84,6 @@ namespace Opm
     class WellTestConfig;
 
     namespace RestartIO { struct RstState; }
-
-
-    struct ScheduleStatic {
-        std::shared_ptr<const Python> m_python_handle;
-        std::string m_input_path;
-        ScheduleRestartInfo rst_info;
-        MessageLimits m_deck_message_limits;
-        UnitSystem m_unit_system;
-        Runspec m_runspec;
-        RSTConfig rst_config;
-        std::optional<int> output_interval;
-        double sumthin{-1.0};
-        bool rptonly{false};
-        bool gaslift_opt_active{false};
-        std::optional<OilVaporizationProperties> oilVap;
-
-        ScheduleStatic() = default;
-
-        explicit ScheduleStatic(std::shared_ptr<const Python> python_handle) :
-            m_python_handle(python_handle)
-        {}
-
-        ScheduleStatic(std::shared_ptr<const Python> python_handle,
-                       const ScheduleRestartInfo& restart_info,
-                       const Deck& deck,
-                       const Runspec& runspec,
-                       const std::optional<int>& output_interval_,
-                       const ParseContext& parseContext,
-                       ErrorGuard& errors);
-
-        template<class Serializer>
-        void serializeOp(Serializer& serializer)
-        {
-            serializer(m_deck_message_limits);
-            serializer(this->rst_info);
-            serializer(m_runspec);
-            serializer(m_unit_system);
-            serializer(this->m_input_path);
-            serializer(rst_info);
-            serializer(rst_config);
-            serializer(this->output_interval);
-            serializer(this->gaslift_opt_active);
-        }
-
-
-        static ScheduleStatic serializationTestObject();
-
-        bool operator==(const ScheduleStatic& other) const;
-    };
-
 
     class Schedule {
     public:
