+++ src/opm/input/eclipse/Schedule/Schedule.cpp
@@ -79,10 +79,6 @@
 #include <opm/input/eclipse/Parser/ParserKeywords/B.hpp>
 #include <opm/input/eclipse/Parser/ParserKeywords/C.hpp>
 #include <opm/input/eclipse/Parser/ParserKeywords/E.hpp>
-#include <opm/input/eclipse/Parser/ParserKeywords/L.hpp>
-#include <opm/input/eclipse/Parser/ParserKeywords/P.hpp>
-#include <opm/input/eclipse/Parser/ParserKeywords/R.hpp>
-#include <opm/input/eclipse/Parser/ParserKeywords/S.hpp>
 #include <opm/input/eclipse/Parser/ParserKeywords/W.hpp>
 
 #include <opm/input/eclipse/Deck/Deck.hpp>
@@ -118,95 +114,9 @@ namespace {
         return Opm::shmatch(pattern, name);
     }
 
-    double sumthin_summary_section(const Opm::SUMMARYSection& section) {
-        const auto entries = section.getKeywordList<Opm::ParserKeywords::SUMTHIN>();
-
-        // Care only about the last SUMTHIN entry in the SUMMARY
-        // section if keyword is present here at all.
-        return entries.empty()
-            ? -1.0 // (<= 0.0)
-            : entries.back()->getRecord(0).getItem(0).getSIDouble(0);
-    }
-
-    bool rptonly_summary_section(const Opm::SUMMARYSection& section) {
-        auto rptonly = false;
-
-        using On = Opm::ParserKeywords::RPTONLY;
-        using Off = Opm::ParserKeywords::RPTONLYO;
-
-        // Last on/off keyword entry "wins".
-        for (const auto& keyword : section) {
-            if (keyword.is<On>())
-                rptonly = true;
-            else if (keyword.is<Off>())
-                rptonly = false;
-        }
-
-        return rptonly;
-    }
-
-    std::optional<Opm::OilVaporizationProperties>
-    vappars_solution_section(const Opm::SOLUTIONSection& section, const int numpvt) {
-        if (section.hasKeyword("VAPPARS")) {
-            const auto& record = section.getKeyword("VAPPARS").getRecord(0);
-            Opm::OilVaporizationProperties ovp(numpvt);
-            double vap1 = record.getItem("OIL_VAP_PROPENSITY").get< double >(0);
-            double vap2 = record.getItem("OIL_DENSITY_PROPENSITY").get< double >(0);
-            Opm::OilVaporizationProperties::updateVAPPARS(ovp, vap1, vap2);
-            return ovp;
-        }
-        return std::nullopt;
-    }
 }
 
 namespace Opm {
-
-    ScheduleStatic::ScheduleStatic(std::shared_ptr<const Python> python_handle,
-                                   const ScheduleRestartInfo& restart_info,
-                                   const Deck& deck,
-                                   const Runspec& runspec,
-                                   const std::optional<int>& output_interval_,
-                                   const ParseContext& parseContext,
-                                   ErrorGuard& errors) :
-        m_python_handle(python_handle),
-        m_input_path(deck.getInputPath()),
-        rst_info(restart_info),
-        m_deck_message_limits( deck ),
-        m_unit_system( deck.getActiveUnitSystem() ),
-        m_runspec( runspec ),
-        rst_config( SOLUTIONSection(deck), parseContext, errors ),
-        output_interval(output_interval_),
-        sumthin(sumthin_summary_section(SUMMARYSection{ deck })),
-        rptonly(rptonly_summary_section(SUMMARYSection{ deck })),
-        gaslift_opt_active(deck.hasKeyword<ParserKeywords::LIFTOPT>()),
-        oilVap(vappars_solution_section(SOLUTIONSection{deck},
-                                        runspec.tabdims().getNumPVTTables()))
-    {
-    }
-
-    ScheduleStatic ScheduleStatic::serializationTestObject()
-    {
-        auto python = std::make_shared<Python>(Python::Enable::OFF);
-        ScheduleStatic st(python);
-        st.m_deck_message_limits = MessageLimits::serializationTestObject();
-        st.m_runspec = Runspec::serializationTestObject();
-        st.m_unit_system = UnitSystem::newFIELD();
-        st.m_input_path = "Some/funny/path";
-        st.rst_config = RSTConfig::serializationTestObject();
-        st.rst_info = ScheduleRestartInfo::serializationTestObject();
-        return st;
-    }
-
-    bool ScheduleStatic::operator==(const ScheduleStatic& other) const
-    {
-        return this->m_input_path == other.m_input_path &&
-               this->m_deck_message_limits == other.m_deck_message_limits &&
-               this->m_unit_system == other.m_unit_system &&
-               this->rst_config == other.rst_config &&
-               this->rst_info == other.rst_info &&
-               this->gaslift_opt_active == other.gaslift_opt_active &&
-               this->m_runspec == other.m_runspec;
-    }
 
     Schedule::Schedule( const Deck& deck,
                         const EclipseGrid& ecl_grid,
